+++ grand/backends/_dataframe.py
@@ -33,10 +33,12 @@ class DataFrameBackend(Backend):
                 use as the node ID
         """
         self._directed = directed
-        self._edge_df = edge_df or pd.DataFrame(
-            columns=[edge_df_source_column, edge_df_target_column]
+        self._edge_df = (
+            edge_df
+            if edge_df is not None
+            else pd.DataFrame(columns=[edge_df_source_column, edge_df_target_column])
         )
-        self._node_df = node_df or None
+        self._node_df = node_df if node_df is not None else None
         self._edge_df_source_column = edge_df_source_column
         self._edge_df_target_column = edge_df_target_column
         self._node_df_id_column = node_df_id_column
@@ -120,11 +122,13 @@ class DataFrameBackend(Backend):
         if self._node_df is not None:
             return [
                 (
-                    node_id,
-                    row.to_dict(),
+                    (
+                        node_id,
+                        row.to_dict(),
+                    )
+                    if include_metadata
+                    else node_id
                 )
-                if include_metadata
-                else node_id
                 for node_id, row in self._node_df.iterrows()
             ]
 
@@ -287,14 +291,12 @@ class DataFrameBackend(Backend):
 
         """
         if self._directed:
-            return (
-                self._edge_df[
-                    (self._edge_df[self._edge_df_source_column] == u)
-                    & (self._edge_df[self._edge_df_target_column] == v)
-                ]
-                .iloc[0]
-                .to_dict()
-            )
+            result = self._edge_df[
+                (self._edge_df[self._edge_df_source_column] == u)
+                & (self._edge_df[self._edge_df_target_column] == v)
+            ]
+            if len(result):
+                return self._edge_as_dict(result.iloc[0])
 
         else:
             left = self._edge_df[
@@ -355,9 +357,11 @@ class DataFrameBackend(Backend):
         else:
             return iter(
                 [
-                    row[self._edge_df_source_column]
-                    if row[self._edge_df_source_column] != u
-                    else row[self._edge_df_target_column]
+                    (
+                        row[self._edge_df_source_column]
+                        if row[self._edge_df_source_column] != u
+                        else row[self._edge_df_target_column]
+                    )
                     for _, row in self._edge_df[
                         (self._edge_df[self._edge_df_source_column] == u)
                         | (self._edge_df[self._edge_df_target_column] == u)
@@ -430,9 +434,11 @@ class DataFrameBackend(Backend):
         else:
             return iter(
                 [
-                    row[self._edge_df_source_column]
-                    if row[self._edge_df_target_column] != u
-                    else row[self._edge_df_target_column]
+                    (
+                        row[self._edge_df_source_column]
+                        if row[self._edge_df_target_column] != u
+                        else row[self._edge_df_target_column]
+                    )
                     for _, row in self._edge_df[
                         (self._edge_df[self._edge_df_target_column] == u)
                         | (self._edge_df[self._edge_df_source_column] == u)
@@ -459,6 +465,19 @@ class DataFrameBackend(Backend):
                 set(self._edge_df[self._edge_df_target_column])
             )
         )
+
+    def get_edge_count(self) -> int:
+        """
+        Get an integer count of the number of edges in this graph.
+
+        Arguments:
+            None
+
+        Returns:
+            int: The count of edges
+
+        """
+        return len(self._edge_df)
 
     def ingest_from_edgelist_dataframe(
         self, edgelist: pd.DataFrame, source_column: str, target_column: str