+++ grand/backends/_sqlbackend.py
@@ -1,11 +1,10 @@
-from typing import Hashable, Generator, Optional, Iterable
+from typing import Hashable, Generator
 import time
 
 import pandas as pd
 import sqlalchemy
-from sqlalchemy.pool import NullPool
-from sqlalchemy.sql import select
-from sqlalchemy import and_, or_, func
+from sqlalchemy.sql import delete, select
+from sqlalchemy import or_, func, Index
 
 from .backend import Backend
 
@@ -60,51 +59,48 @@ class SQLBackend(Backend):
         self._connection = self._engine.connect()
         self._metadata = sqlalchemy.MetaData()
 
-        if not self._engine.dialect.has_table(self._connection, self._node_table_name):
-            self._node_table = sqlalchemy.Table(
-                self._node_table_name,
-                self._metadata,
-                sqlalchemy.Column(
-                    self._primary_key,
-                    sqlalchemy.String(_DEFAULT_SQL_STR_LEN),
-                    primary_key=True,
-                ),
-                sqlalchemy.Column("_metadata", sqlalchemy.JSON),
-            )
-            self._node_table.create(self._engine)
-        else:
-            self._node_table = sqlalchemy.Table(
-                self._node_table_name,
-                self._metadata,
-                autoload=True,
-                autoload_with=self._engine,
-            )
+        # Create nodes table
+        self._node_table = sqlalchemy.Table(
+            self._node_table_name,
+            self._metadata,
+            sqlalchemy.Column(
+                self._primary_key,
+                sqlalchemy.String(_DEFAULT_SQL_STR_LEN),
+                primary_key=True,
+            ),
+            sqlalchemy.Column("_metadata", sqlalchemy.JSON),
+        )
+        self._node_table.create(self._engine, checkfirst=True)
 
-        if not self._engine.dialect.has_table(self._connection, self._edge_table_name):
-            self._edge_table = sqlalchemy.Table(
-                self._edge_table_name,
-                self._metadata,
-                sqlalchemy.Column(
-                    self._primary_key,
-                    sqlalchemy.String(_DEFAULT_SQL_STR_LEN),
-                    primary_key=True,
-                ),
-                sqlalchemy.Column("_metadata", sqlalchemy.JSON),
-                sqlalchemy.Column(
-                    self._edge_source_key, sqlalchemy.String(_DEFAULT_SQL_STR_LEN)
-                ),
-                sqlalchemy.Column(
-                    self._edge_target_key, sqlalchemy.String(_DEFAULT_SQL_STR_LEN)
-                ),
-            )
-            self._edge_table.create(self._engine)
-        else:
-            self._edge_table = sqlalchemy.Table(
-                self._edge_table_name,
-                self._metadata,
-                autoload=True,
-                autoload_with=self._engine,
-            )
+        source_column = sqlalchemy.Column(
+            self._edge_source_key, sqlalchemy.String(_DEFAULT_SQL_STR_LEN)
+        )
+
+        target_column = sqlalchemy.Column(
+            self._edge_target_key, sqlalchemy.String(_DEFAULT_SQL_STR_LEN)
+        )
+
+        # Create edges table
+        self._edge_table = sqlalchemy.Table(
+            self._edge_table_name,
+            self._metadata,
+            sqlalchemy.Column(
+                self._primary_key,
+                sqlalchemy.String(_DEFAULT_SQL_STR_LEN),
+                primary_key=True,
+            ),
+            sqlalchemy.Column("_metadata", sqlalchemy.JSON),
+            source_column,
+            target_column,
+        )
+        self._edge_table.create(self._engine, checkfirst=True)
+
+        # Create source and target index
+        sindex = Index("edge_source", source_column)
+        sindex.create(self._engine, checkfirst=True)
+
+        tindex = Index("edge_target", target_column)
+        tindex.create(self._engine, checkfirst=True)
 
     def is_directed(self) -> bool:
         """
@@ -147,16 +143,27 @@ class SQLBackend(Backend):
             existing_metadata.update(metadata)
             self._connection.execute(
                 self._node_table.update().where(
-                    self._node_table.c[self._primary_key] == node_name
+                    self._node_table.c[self._primary_key] == str(node_name)
                 ),
-                **{"_metadata": existing_metadata},
+                parameters={"_metadata": existing_metadata},
             )
         else:
             self._connection.execute(
                 self._node_table.insert(),
-                **{self._primary_key: node_name, "_metadata": metadata},
+                parameters={self._primary_key: node_name, "_metadata": metadata},
             )
         return node_name
+
+    def add_nodes_from(self, nodes_for_adding, **attr):
+        nodes = [
+            {
+                self._primary_key: node,
+                "_metadata": {**attr, **metadata},
+            }
+            for node, metadata in nodes_for_adding
+        ]
+
+        self._connection.execute(self._node_table.insert(), nodes)
 
     def _upsert_node(self, node_name: Hashable, metadata: dict) -> Hashable:
         """
@@ -174,15 +181,38 @@ class SQLBackend(Backend):
         if node_exists:
             self._connection.execute(
                 self._node_table.update().where(
-                    self._node_table.c[self._primary_key] == node_name
+                    self._node_table.c[self._primary_key] == str(node_name)
                 ),
-                **{"_metadata": metadata},
+                parameters={"_metadata": metadata},
             )
         else:
             self._connection.execute(
                 self._node_table.insert(),
-                **{self._primary_key: node_name, "_metadata": metadata},
+                parameters={self._primary_key: node_name, "_metadata": metadata},
             )
+
+    def remove_node(self, u: Hashable) -> None:
+        """
+        Removes nodes and related edges for name.
+
+        Args:
+            u (Hashable): id of the node
+        """
+
+        # Remove nodes
+        statement = delete(self._node_table).where(
+            self._node_table.c[self._primary_key] == str(u)
+        )
+        self._connection.execute(statement)
+
+        # Remove edges for node
+        statement = delete(self._edge_table).where(
+            or_(
+                self._edge_table.c[self._edge_source_key] == str(u),
+                self._edge_table.c[self._edge_target_key] == str(u)
+            )
+        )
+        self._connection.execute(statement)
 
     def all_nodes_as_iterable(self, include_metadata: bool = False) -> Generator:
         """
@@ -196,10 +226,18 @@ class SQLBackend(Backend):
             Generator: A generator of all nodes (arbitrary sort)
 
         """
-        results = self._connection.execute(self._node_table.select()).fetchall()
         if include_metadata:
-            return [(row[self._primary_key], row["_metadata"]) for row in results]
-        return [row[self._primary_key] for row in results]
+            sql = self._node_table.select()
+        else:
+            sql = self._node_table.select().with_only_columns(
+                self._node_table.c[self._primary_key]
+            )
+
+        results = []
+        for x in self._connection.execute(sql):
+            results.append(x if include_metadata else x[0])
+
+        return results
 
     def has_node(self, u: Hashable) -> bool:
         """
@@ -214,10 +252,10 @@ class SQLBackend(Backend):
         return len(
             self._connection.execute(
                 self._node_table.select().where(
-                    self._node_table.c[self._primary_key] == u
+                    self._node_table.c[self._primary_key] == str(u)
                 )
             ).fetchall()
-        )
+        ) > 0
 
     def add_edge(self, u: Hashable, v: Hashable, metadata: dict):
         """
@@ -245,7 +283,7 @@ class SQLBackend(Backend):
         try:
             self._connection.execute(
                 self._edge_table.insert(),
-                **{
+                parameters={
                     self._primary_key: pk,
                     self._edge_source_key: u,
                     self._edge_target_key: v,
@@ -260,10 +298,23 @@ class SQLBackend(Backend):
                 self._edge_table.update().where(
                     self._edge_table.c[self._primary_key] == pk
                 ),
-                **{"_metadata": existing_metadata},
+                parameters={"_metadata": existing_metadata},
             )
 
         return pk
+
+    def add_edges_from(self, ebunch_to_add, **attr):
+        edges = [
+            {
+                self._primary_key: f"__{u}__{v}",
+                self._edge_source_key: u,
+                self._edge_target_key: v,
+                "_metadata": {**attr, **metadata},
+            }
+            for u, v, metadata in ebunch_to_add
+        ]
+
+        self._connection.execute(self._edge_table.insert(), edges)
 
     def all_edges_as_iterable(self, include_metadata: bool = False) -> Generator:
         """
@@ -274,16 +325,18 @@ class SQLBackend(Backend):
 
         Returns:
             Generator: A generator of all edges (arbitrary sort)
-
         """
-        return iter(
-            [
-                (e.Source, e.Target, e._metadata)
-                if include_metadata
-                else (e.Source, e.Target)
-                for e in self._connection.execute(self._edge_table.select()).fetchall()
-            ]
-        )
+
+        columns = [
+            self._edge_table.c[self._edge_source_key],
+            self._edge_table.c[self._edge_target_key],
+        ]
+
+        if include_metadata:
+            columns.append(self._edge_table.c["_metadata"])
+
+        sql = self._edge_table.select().with_only_columns(*columns)
+        return self._connection.execute(sql).fetchall()
 
     def get_node_by_id(self, node_name: Hashable):
         """
@@ -296,16 +349,17 @@ class SQLBackend(Backend):
             dict: The metadata associated with this node
 
         """
-        res = (
-            self._connection.execute(
-                self._node_table.select().where(
-                    self._node_table.c[self._primary_key] == node_name
-                )
+
+        res = self._connection.execute(
+            self._node_table.select().where(
+                self._node_table.c[self._primary_key] == str(node_name)
             )
-            .fetchone()
-            ._metadata
-        )
-        return res
+        ).fetchone()
+
+        if res:
+            return res._metadata
+
+        raise KeyError(f"Node {node_name} not found")
 
     def get_edge_by_id(self, u: Hashable, v: Hashable):
         """
@@ -321,28 +375,26 @@ class SQLBackend(Backend):
         """
         if self._directed:
             pk = f"__{u}__{v}"
-            return (
-                self._connection.execute(
-                    self._edge_table.select().where(
-                        self._edge_table.c[self._primary_key] == pk
-                    )
+            result = self._connection.execute(
+                self._edge_table.select().where(
+                    self._edge_table.c[self._primary_key] == pk
                 )
-                .fetchone()
-                ._metadata
-            )
+            ).fetchone()
+            if result:
+                return result._metadata
+            raise KeyError(f"Edge {u}-{v} not found.")
         else:
-            return (
-                self._connection.execute(
-                    self._edge_table.select().where(
-                        or_(
-                            (self._edge_table.c[self._primary_key] == f"__{u}__{v}"),
-                            (self._edge_table.c[self._primary_key] == f"__{v}__{u}"),
-                        )
+            result = self._connection.execute(
+                self._edge_table.select().where(
+                    or_(
+                        (self._edge_table.c[self._primary_key] == f"__{u}__{v}"),
+                        (self._edge_table.c[self._primary_key] == f"__{v}__{u}"),
                     )
                 )
-                .fetchone()
-                ._metadata
-            )
+            ).fetchone()
+            if result:
+                return result._metadata
+            raise KeyError(f"Edge {u}-{v} not found.")
 
     def get_node_neighbors(
         self, u: Hashable, include_metadata: bool = False
@@ -357,27 +409,32 @@ class SQLBackend(Backend):
             Generator
 
         """
+
         if self._directed:
             res = self._connection.execute(
-                self._edge_table.select().where(
-                    self._edge_table.c[self._edge_source_key] == u
-                )
+                self._edge_table.select()
+                .where(self._edge_table.c[self._edge_source_key] == str(u))
+                .order_by(self._edge_table.c[self._primary_key])
             ).fetchall()
         else:
             res = self._connection.execute(
-                self._edge_table.select().where(
+                self._edge_table.select()
+                .where(
                     or_(
-                        (self._edge_table.c[self._edge_source_key] == u),
-                        (self._edge_table.c[self._edge_target_key] == u),
+                        (self._edge_table.c[self._edge_source_key] == str(u)),
+                        (self._edge_table.c[self._edge_target_key] == str(u)),
                     )
                 )
+                .order_by(self._edge_table.c[self._primary_key])
             ).fetchall()
+
+        res = [x._asdict() for x in res]
 
         if include_metadata:
             return {
                 (
                     r[self._edge_source_key]
-                    if r[self._edge_source_key] != u
+                    if r[self._edge_source_key] != str(u)
                     else r[self._edge_target_key]
                 ): r["_metadata"]
                 for r in res
@@ -387,7 +444,7 @@ class SQLBackend(Backend):
             [
                 (
                     r[self._edge_source_key]
-                    if r[self._edge_source_key] != u
+                    if r[self._edge_source_key] != str(u)
                     else r[self._edge_target_key]
                 )
                 for r in res
@@ -409,25 +466,29 @@ class SQLBackend(Backend):
         """
         if self._directed:
             res = self._connection.execute(
-                self._edge_table.select().where(
-                    self._edge_table.c[self._edge_target_key] == u
-                )
+                self._edge_table.select()
+                .where(self._edge_table.c[self._edge_target_key] == str(u))
+                .order_by(self._edge_table.c[self._primary_key])
             ).fetchall()
         else:
             res = self._connection.execute(
-                self._edge_table.select().where(
+                self._edge_table.select()
+                .where(
                     or_(
-                        (self._edge_table.c[self._edge_target_key] == u),
-                        (self._edge_table.c[self._edge_source_key] == u),
+                        (self._edge_table.c[self._edge_target_key] == str(u)),
+                        (self._edge_table.c[self._edge_source_key] == str(u)),
                     )
                 )
+                .order_by(self._edge_table.c[self._primary_key])
             ).fetchall()
+
+        res = [x._asdict() for x in res]
 
         if include_metadata:
             return {
                 (
                     r[self._edge_source_key]
-                    if r[self._edge_source_key] != u
+                    if r[self._edge_source_key] != str(u)
                     else r[self._edge_target_key]
                 ): r["_metadata"]
                 for r in res
@@ -437,14 +498,14 @@ class SQLBackend(Backend):
             [
                 (
                     r[self._edge_source_key]
-                    if r[self._edge_source_key] != u
+                    if r[self._edge_source_key] != str(u)
                     else r[self._edge_target_key]
                 )
                 for r in res
             ]
         )
 
-    def get_node_count(self) -> Iterable:
+    def get_node_count(self) -> int:
         """
         Get an integer count of the number of nodes in this graph.
 
@@ -456,7 +517,22 @@ class SQLBackend(Backend):
 
         """
         return self._connection.execute(
-            select([func.count()]).select_from(self._node_table)
+            select(func.count()).select_from(self._node_table)
+        ).scalar()
+
+    def get_edge_count(self) -> int:
+        """
+        Get an integer count of the number of edges in this graph.
+
+        Arguments:
+            None
+
+        Returns:
+            int: The count of edges
+
+        """
+        return self._connection.execute(
+            select(func.count()).select_from(self._edge_table)
         ).scalar()
 
     def out_degrees(self, nbunch=None):
@@ -474,20 +550,22 @@ class SQLBackend(Backend):
         if nbunch is None:
             where_clause = None
         elif isinstance(nbunch, (list, tuple)):
-            where_clause = self._edge_table.c[self._edge_source_key].in_(nbunch)
+            where_clause = self._edge_table.c[self._edge_source_key].in_(
+                [str(x) for x in nbunch]
+            )
         else:
             # single node:
-            where_clause = self._edge_table.c[self._edge_source_key] == nbunch
+            where_clause = self._edge_table.c[self._edge_source_key] == str(nbunch)
 
         if self._directed:
             query = (
-                select([self._edge_table.c[self._edge_source_key], func.count()])
+                select(self._edge_table.c[self._edge_source_key], func.count())
                 .select_from(self._edge_table)
                 .group_by(self._edge_table.c[self._edge_source_key])
             )
         else:
             query = (
-                select([self._edge_table.c[self._edge_source_key], func.count()])
+                select(self._edge_table.c[self._edge_source_key], func.count())
                 .select_from(self._edge_table)
                 .group_by(self._edge_table.c[self._edge_source_key])
             )
@@ -495,10 +573,7 @@ class SQLBackend(Backend):
         if where_clause is not None:
             query = query.where(where_clause)
 
-        results = {
-            r[self._edge_source_key]: r[1]
-            for r in self._connection.execute(query).fetchall()
-        }
+        results = {r[0]: r[1] for r in self._connection.execute(query)}
 
         if nbunch and not isinstance(nbunch, (list, tuple)):
             return results.get(nbunch, 0)
@@ -519,20 +594,22 @@ class SQLBackend(Backend):
         if nbunch is None:
             where_clause = None
         elif isinstance(nbunch, (list, tuple)):
-            where_clause = self._edge_table.c[self._edge_target_key].in_(nbunch)
+            where_clause = self._edge_table.c[self._edge_target_key].in_(
+                [str(x) for x in nbunch]
+            )
         else:
             # single node:
-            where_clause = self._edge_table.c[self._edge_target_key] == nbunch
+            where_clause = self._edge_table.c[self._edge_target_key] == str(nbunch)
 
         if self._directed:
             query = (
-                select([self._edge_table.c[self._edge_target_key], func.count()])
+                select(self._edge_table.c[self._edge_target_key], func.count())
                 .select_from(self._edge_table)
                 .group_by(self._edge_table.c[self._edge_target_key])
             )
         else:
             query = (
-                select([self._edge_table.c[self._edge_target_key], func.count()])
+                select(self._edge_table.c[self._edge_target_key], func.count())
                 .select_from(self._edge_table)
                 .group_by(self._edge_table.c[self._edge_target_key])
             )
@@ -540,10 +617,7 @@ class SQLBackend(Backend):
         if where_clause is not None:
             query = query.where(where_clause)
 
-        results = {
-            r[self._edge_target_key]: r[1]
-            for r in self._connection.execute(query).fetchall()
-        }
+        results = {r[0]: r[1] for r in self._connection.execute(query)}
 
         if nbunch and not isinstance(nbunch, (list, tuple)):
             return results.get(nbunch, 0)
@@ -619,3 +693,9 @@ class SQLBackend(Backend):
             "edge_count": len(edgelist),
             "edge_duration": edge_toc,
         }
+
+    def commit(self):
+        self._connection.commit()
+
+    def close(self):
+        self._connection.close()