+++ CHANGELOG.md
@@ -1,11 +1,51 @@
 # CHANGELOG
 
-## **0.4.2** (Unreleased)
+## **0.6.0** (December 8, 2024)
+
+-   Features:
+    -   Upgrade codspeed CI runner to v3
+-   Bugfixes:
+    -   Fix `get_node_by_id` method in `SQLBackend` (#59. thanks @davidmezzetti!)
+    -   Add `remove_node` to the `CachedBackend` (#59. thanks @davidmezzetti!)
+
+## **0.5.3** (December 7, 2024)
+
+-   Features:
+    -   Add `remove_node` method to SQLBackend (#58. thanks @davidmezzetti!)
+-   Bugfixes:
+    -   Modify sqlite backend `has_node` to return a bool vs int (#58. thanks @davidmezzetti!)
+    -   Update method parameters to conform to existing naming standards (#58. thanks @davidmezzetti!)
+
+## **0.5.2** (June 4, 2024)
+
+-   Bugfixes:
+    -   Fixes SQLBackend mistakenly referencing Nodes table for enumeration of Edges (#55, #56)
+
+## **0.5.1** (May 13, 2024)
+
+-   Bugfixes:
+    -   Fixes SQLBackend bug where graphs would not commit down to disk after transactions (#51, thanks @acthecoder23!)
+
+## **0.5.0** (April 17, 2024)
+
+-   Features
+    -   https://github.com/aplbrain/grand/pull/45 (thanks @davidmezzetti!)
+        -   Improved support for SQL backend, including an index on edge.sources and edge.targets
+        -   Improved batch-adding performance for nodes and edges
+-   Housekeeping
+    -   Removed extra arguments to Graph constructor and improved Graph kwargs handing
+    -   Added `grand.DiGraph` convenience wrapper for directed graphs
+-   Dialects
+    -   Expanded Networkit dialect test coverage
+    -   Added support for exporting NetworkX graphs by adding `graph` attribute
+
+## **0.4.2** (May 7, 2022)
 
 -   Backends
     -   NEW: `DataFrameBackend` supports operations on a pandas-like API.
 -   Housekeeping
     -   Added tests for metadata stores
+    -   Added IGraph and Networkit backends to the standard GitHub Actions CI test suite
 
 ## **0.4.1** (February 10, 2022)